Move main screen to seperate file
